// src/types.d.ts
export declare module "@mui/material/styles" {
  interface PaletteOptions {
    lighterGray: string;
  }
  interface Palette {
    lighterGray: string;
  }
}

export interface fakeStoreProduct {
  id: number;
  title: string;
  price: string;
  category: string;
  description: string;
  image: string;
}